+++ src/components/deprecated/index.ts
@@ -12,8 +12,6 @@ export { default as SSidebarNavChild } from './SSidebarNavChild.vue';
 export { default as SSidebarNavGroup } from './SSidebarNavGroup.vue';
 export { default as SSidebarNavParent } from './SSidebarNavParent.vue';
 export { default as SDescriptionList } from './SDescriptionList.vue';
-export { default as SSectionTitle } from './SSectionTitle.vue';
-export { default as SSectionDescription } from './SSectionDescription.vue';
 export { default as STabsLine } from './STabsLine.vue';
 export * from './sTables';
 export { SInputAlt, SInputDate, SInputDropdown, SInputError, SInputHelper, SInputMask, SInputRadio } from './sInputs';