+++ src/components/spartan/SSectionDescription/index.ts
@@ -1,0 +1,1 @@
+export { default as SSectionDescription } from './SSectionDescription.vue';